+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-11-12
+
+### Added
+
+- Display: Add Display output
+- DisplayElement: 3 types of display elements
+  - DisplayElementText to draw a text
+  - DisplayElementTexture to draw a texture from atlas
+  - DisplayElementDiv to draw a colored rectangle
+
 ## [0.3.2] - 2024-11-05
 
 ### Changed